Developed version 0.4

+++ squirrel.frontier/src/main/java/org/dice_research/squirrel/data/uri/filter/MongoDBKnowUriFilter.java
@@ -13,6 +13,7 @@ import java.util.Set;
 
 import org.bson.Document;
 import org.bson.conversions.Bson;
+import org.dice_research.squirrel.configurator.MongoConfiguration;
 import org.dice_research.squirrel.data.uri.CrawleableUri;
 import org.dice_research.squirrel.data.uri.UriType;
 import org.dice_research.squirrel.deduplication.hashing.HashValue;
@@ -23,12 +24,23 @@ import org.slf4j.LoggerFactory;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.MongoClient;
+import com.mongodb.MongoClientOptions;
+import com.mongodb.ServerAddress;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Indexes;
 
+/**
+ * 
+ * Filter implementation for use with MongoDB
+ * 
+ * * @author Geraldo Souza Junior (gsjunior@mail.uni-paderborn.de)
+ *
+ */
+
+@SuppressWarnings("deprecation")
 public class MongoDBKnowUriFilter implements KnownUriFilter, Cloneable, Closeable,UriHashCustodian {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MongoDBKnowUriFilter.class);
@@ -46,15 +58,32 @@ public class MongoDBKnowUriFilter implements KnownUriFilter, Cloneable, Closeabl
     public static final String COLUMN_IP = "ipAddress";
     public static final String COLUMN_TYPE = "type";
     public static final String COLUMN_HASH_VALUE = "hashValue";
+    private static final boolean PERSIST = System.getenv("QUEUE_FILTER_PERSIST") == null ? false : Boolean.parseBoolean(System.getenv("QUEUE_FILTER_PERSIST"));
     /**
      * Used as a default hash value for URIS, will be replaced by real hash value as soon as it has been computed.
      */
     private static final String DUMMY_HASH_VALUE = "dummyValue";
 
     public MongoDBKnowUriFilter(String hostName, Integer port) {
-    
+ 
+        LOGGER.info("Filter Persistance: " + PERSIST);
+
     	
-        client = new MongoClient(hostName, port);
+    	MongoClientOptions.Builder optionsBuilder = MongoClientOptions.builder();
+        MongoConfiguration mongoConfiguration = MongoConfiguration.getMDBConfiguration();
+	
+		if(mongoConfiguration != null &&(mongoConfiguration.getConnectionTimeout() != null && mongoConfiguration.getSocketTimeout() != null && mongoConfiguration.getServerTimeout() != null)) {
+			optionsBuilder.connectTimeout(mongoConfiguration.getConnectionTimeout());
+			optionsBuilder.socketTimeout(mongoConfiguration.getSocketTimeout());
+			optionsBuilder.serverSelectionTimeout(mongoConfiguration.getServerTimeout());
+			
+			MongoClientOptions options = optionsBuilder.build();
+
+			client = new MongoClient(new ServerAddress(hostName, port),options);
+			
+		}else {
+			client = new MongoClient(hostName, port);
+		}
     }
 
     @Override
@@ -87,18 +116,18 @@ public class MongoDBKnowUriFilter implements KnownUriFilter, Cloneable, Closeabl
 
     public Document crawleableUriToMongoDocument(CrawleableUri uri) {
 
-        InetAddress ipAddress = uri.getIpAddress();
-        @SuppressWarnings("deprecation")
         UriType uriType = uri.getType();
 
-        return new Document("ipAddress", ipAddress.toString()).append("type", uriType.toString()).append("uri",
-                uri.getUri().toString());
+        return new Document("uri", uri.getUri().toString()).append("type", uriType.toString());
 
     }
 
     @Override
     public void close() throws IOException {
-        mongoDB.getCollection(COLLECTION_NAME).drop();
+        if(!PERSIST) {
+            mongoDB.getCollection(COLLECTION_NAME).drop();
+
+        }
         client.close();
     }
 
@@ -136,10 +165,7 @@ public class MongoDBKnowUriFilter implements KnownUriFilter, Cloneable, Closeabl
     
     @Override
     public void addHashValuesForUris(List<CrawleableUri> uris) {
-        for (CrawleableUri uri : uris) {
-//            r.db(DATABASE_NAME).table(TABLE_NAME).filter(doc -> doc.getField(COLUMN_URI).eq(uri.getUri().toString())).
-//                update(r.hashMap(COLUMN_HASH_VALUE, ((HashValue) uri.getData(Constants.URI_HASH_KEY)).encodeToString())).run(connector.connection);
-        }
+
     }
